move authentication logic to auth.ts
