+++ server/src/auth.ts
@@ -1,13 +1,18 @@
-import { Request, Response } from 'express';
 import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
 import crypto from 'crypto';
 import nodemailer from 'nodemailer';
 import dotenv from 'dotenv';
 import { Database } from 'sqlite';
+import { Request, Response, NextFunction } from 'express';
+import { ObjectHandler } from './ObjectHandler';
+
+
 
 dotenv.config();
 
+
+const secret = process.env.JWT_SECRET || 'your_jwt_secret';
 
 export const register = async (req: Request, res: Response, db: any) => {
   const { name, email, password } = req.body;
@@ -92,6 +97,32 @@ export const login = async (req: Request, res: Response, db: Database) => {
   }
 };
 
+export const checkOwnership = (db: Database, oh: ObjectHandler) => {
+    return async (req: Request, res: Response, next: NextFunction) => {
+        const token = req.header('Authorization')?.replace('Bearer ', '');
+        if (!token) {
+            return res.status(401).json({ message: 'Authentication required' });
+        }
+
+        try {
+            const decoded = jwt.verify(token, secret) as { id: string; email: string };
+            const userFromTokenId = await oh.getUser(decoded.id, db);
+            const userFromParamsId = await oh.getUserByMail(req.body.email, db);
+
+            if(!userFromTokenId || !userFromParamsId) {
+                return res.status(404).json({ message: 'User not found' });
+            }
+
+            if (userFromTokenId?.name !== "admin" && userFromParamsId?.name !== userFromTokenId?.name) {
+                return res.status(403).json({ message: 'Forbidden: You can only edit your own data' });
+            }
+
+            next();
+        } catch (error) {
+            return res.status(401).json({ message: 'Invalid token' });
+        }
+    };
+};
 
 const sendPasswordResetEmail = async (email: string, token: string) => {
